+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - actuation: Extend ImuData with raw measurements
 - actuation: Log raw IMU measurements from pi3hat interface
 - envs: Start reward submodule with a wheeled inverted pendulum reward
+- envs: Wheeled inverted pendulum environment for reduced-model testing
 - utils: Raise an exception when trying to configure agent from an interpreter
 
 ### Changed